refactor using bun
