+++ Backend/IBSLoverBackend/src/controllers/toilet/removeToilet.ts
