+++ Backend/IBSLoverBackend/src/controllers/toilet/voteToilet.ts
