+++ Backend/IBSLoverBackend/src/database/models/toilet.model.ts
